% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_files.R
\docType{data}
\name{corpus_files}
\alias{corpus_files}
\title{Example corpus files}
\format{
`corpus_files`
A list of 2:
\describe{
  \item{1}{Vector of sound files.}
  \item{2}{Vector of TextGrid files.}
}
}
\usage{
corpus_files
}
\description{
An example corpus file list created with \code{get_corpus_files}. Can be run with the example data provided on GitHub.
}
\keyword{datasets}
