% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_band_correlations.R
\name{get_band_correlations}
\alias{get_band_correlations}
\title{Compute correlations between chunks of Mel-filterbank energy features}
\usage{
get_band_correlations(band_samples, n_bands = 20, method = "spearman")
}
\arguments{
\item{band_samples}{Samples of chunks of Mel-filterbank energy features. Typically created with \code{get_band_samples}.}

\item{n_bands}{Number of bands that should be sampled. Defaults to \code{20}.}

\item{method}{A character string indicating which correlation coefficient (or covariance) is to be computed. One of \code{"spearman"} (default), \code{"kendall"}, or \code{"pearson"}: can be abbreviated.}
}
\value{
A data frame.
}
\description{
This function creates lists of sound and TextGrid files. This is the initial step of the CFBSF analysis.
}
\examples{
\dontrun{
band_correlations <- get_band_correlations(band_samples)
}

}
\author{
Dominic Schmitz
}
