% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_band_samples.R
\name{get_band_samples}
\alias{get_band_samples}
\title{Sample chunks of Mel-filterbank energy features}
\usage{
get_band_samples(mel_features, n_bands = 20, seed = 16121991)
}
\arguments{
\item{mel_features}{List of Mel-filterbank energy features. Typically created with \code{get_mel}.}

\item{n_bands}{Number of bands that should be sampled. Defaults to \code{20}.}

\item{seed}{Set seed for reproducibility. Defaults to \code{161211991}.}
}
\value{
A data frame.
}
\description{
This function samples \code{n} columns of Mel-filterbank energy features as range, i.e. the order of columns is retained.
}
\examples{
\dontrun{band_samples <- get_band_samples(mel_features = mel_features, n_bands = 20)}

}
\author{
Dominic Schmitz
}
