% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_band_summaries.R
\name{get_band_summaries}
\alias{get_band_summaries}
\title{Compute Mel-based summaries for all chunks of a word}
\usage{
get_band_summaries(
  mel_features,
  chunk_duration,
  n_cols = 20,
  connected = TRUE,
  seed = 16121991,
  method = "spearman",
  progress = TRUE
)
}
\arguments{
\item{mel_features}{List of Mel-filterbank energy features. Typically created with \code{get_mel}.}

\item{chunk_duration}{List of chunk duration. Typically created with \code{get_chunk_duration}.}

\item{n_cols}{Number of columns that should be sampled. Defaults to \code{20}.}

\item{connected}{Should the sampled columns be connected, e.g. columns 1-20, or discontinuous? Defaults to \code{TRUE} (connected).}

\item{seed}{Set seed for reproducibility. Defaults to \code{161211991}.}

\item{method}{A character string indicating which correlation coefficient (or covariance) is to be computed. One of \code{"spearman"} (default), \code{"kendall"}, or \code{"pearson"}: can be abbreviated.}

\item{progress}{Show a console progress bar. Defaults to \code{TRUE}.}
}
\value{
A list object.
}
\description{
This function computes Mel-based summaries for all chunks of a word (or other unit). Uses the outputs of \code{get_mel} and \code{get_chunk_duration}. The length of all summaries is padded with zeros to achieve similar lengths.
}
\examples{
\dontrun{band_summaries <- get_band_summaries(mel_features = my_mel_features, chunk_duration = my_chunk_durations)}

}
\author{
Dominic Schmitz
}
