% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corpus_files.R
\name{get_corpus_files}
\alias{get_corpus_files}
\title{Collect corpus sound files and TextGrids}
\usage{
get_corpus_files(path = NULL, sound_types = "wav")
}
\arguments{
\item{path}{The path to the directory containing sound and TextGrid files.}

\item{sound_types}{The types of sound files that should be listed. Defaults to /code{"wav}.}
}
\value{
A list object.
}
\description{
This function creates lists of sound and TextGrid files. This is the initial step of the CFBS analysis.
}
\examples{
corpus_files <- get_corpus_files(path = "data", "wav")

}
\author{
Dominic Schmitz
}
