% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fbs.R
\name{get_fbs}
\alias{get_fbs}
\title{Compute Mel-filterbank energy features from audio}
\usage{
get_fbs(
  chopped_dir,
  sample_rate = 16000,
  winlen = 0.005,
  steptime = 0.01,
  fft = 512,
  nfilt = 21,
  dither = FALSE
)
}
\arguments{
\item{chopped_dir}{Directory of the individual sound files. Typically created with \code{get_word_sound_files}.}

\item{sample_rate}{Sample rate of your wav files. Default is \code{16000}.}

\item{winlen}{The length of the analysis window in seconds. Default is \code{0.005}.}

\item{steptime}{The step between successive windows in seconds. Default is \code{0.01}.}

\item{fft}{The FFT size. Default is \code{512}.}

\item{nfilt}{The number of filters in the filterbank. Default is \code{21}.}

\item{dither}{Add dither to MEL features. Defaults to \code{FALSE}.}
}
\value{
A list object.
}
\description{
This function computes Mel-filterbank energy features from audio signals. Clone of \code{get_mel}.
}
\examples{
\dontrun{fbs_data <- get_fbs(chopped_dir = "/chopped")}

}
\references{
Ligges, U., Krey, S., Mersmann, O., & Schnackenberg, S. (2023). tuneR: Analysis of Music and Speech. Retrieved from https://CRAN.R-project.org/package=tuneR

Signal developers. (2023). signal: Signal processing. Retrieved from https://r-forge.r-project.org/projects/signal/
}
\author{
Dominic Schmitz
}
