% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mel.R
\name{get_mel}
\alias{get_mel}
\title{Compute Mel-filterbank energy features from audio}
\usage{
get_mel(
  chopped_dir,
  winlen = 0.025,
  winstep = 0.01,
  nfft = 512,
  nfilt = 26,
  lowfreq = 0,
  highfreq = NULL
)
}
\arguments{
\item{chopped_dir}{Directory of the individual sound files. Typically created with \code{get_word_sound_files}.}

\item{winlen}{The length of the analysis window in seconds. Default is \code{0.025}.}

\item{winstep}{The step between successive windows in seconds. Default is \code{0.01}.}

\item{nfft}{The FFT size. Default is \code{512}.}

\item{nfilt}{The number of filters in the filterbank. Default is \code{26}.}

\item{lowfreq}{Lowest band edge of Mel filters. In Hz, default is \code{0}.}

\item{highfreq}{Highest band edge of Mel filters. In Hz, default is \code{samplerate/2}.}
}
\value{
A list object.
}
\description{
This function computes Mel-filterbank energy features from audio signals.
}
\examples{
\dontrun{mel_features <- get_mel(chopped_dir = "/chopped")}

}
\references{
Uwe Ligges, Sebastian Krey, Olaf Mersmann, and Sarah Schnackenberg (2023). tuneR: Analysis of Music and Speech. URL: https://CRAN.R-project.org/package=tuneR

James Lyons, Darren Yow-Bang Wang, Gianluca, Hanan Shteingart, Erik Mavrinac, Yash Gaurkar, Watcharapol Watcharawisetkul, Sam Birch, Lu Zhihe, Josef Hölzl, Janis Lesinskis, Henrik Almér, Chris Lord, & Adam Stark. (2020). jameslyons/python_speech_features: release v0.6.1 (0.6.1). Zenodo. https://doi.org/10.5281/zenodo.3607820
}
\author{
Dominic Schmitz
}
