% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_word_duration.R
\name{get_word_duration}
\alias{get_word_duration}
\title{Get word durations, starts, ends}
\usage{
get_word_duration(tg_files, main_tier, extra_tier = NULL, encoding = "auto")
}
\arguments{
\item{tg_files}{A vector of all TextGrid files you want to include. Typically created with \code{get_corpus_files}.}

\item{main_tier}{The tier with your relevant intervals, e.g. words. Can be given by its name or by its index.}

\item{extra_tier}{Optional: Include another tier. This tier will be used in unique identifier creation. Can be given by its name or by its index.}

\item{encoding}{Defaults to \code{UTF-8}.}
}
\value{
A data frame.
}
\description{
This function collects the start points, end points, and durations of all words (or whatever else is annotated in your relevant TextGrid tier).
}
\examples{
\dontrun{corpus_files <- get_corpus_files(path = "data", "wav")

wd_df <- get_word_duration(tg_files = corpus_files[[2]], main_tier = 1)
wd_df <- get_word_duration(tg_files = corpus_files[[2]], main_tier = 1, extra_tier = 2)

wd_df <- get_word_duration(tg_files = corpus_files[[2]], main_tier = "word")
wd_df <- get_word_duration(tg_files = corpus_files[[2]], main_tier = "word", extra_tier = "ID")}

}
\references{
Bořil, T., & Skarnitzl, R. (2016). Tools rPraat and mPraat. In P. Sojka, A. Horák, I. Kopeček, & K. Pala (Eds.), Text, Speech, and Dialogue (pp. 367-374). Springer International Publishing.
}
\author{
Dominic Schmitz
}
