% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_word_sound_files.R
\name{get_word_sound_files}
\alias{get_word_sound_files}
\title{Cut sound files}
\usage{
get_word_sound_files(
  sound_files,
  word_durations,
  extra_tier = NULL,
  directory,
  fileType = "auto",
  from = 1,
  to = Inf,
  units = "samples"
)
}
\arguments{
\item{sound_files}{A vector of all wav files you want to include.}

\item{word_durations}{The durations of words or other relevant intervals. Typically identified using \code{get_word_duration}.}

\item{extra_tier}{Optional: Include another tier. This tier will be used in unique identifier creation. Can be given by its name or by its index.}

\item{directory}{Specify the directory in which the new wav files should be saved.}

\item{fileType}{Type of audio file: \code{"wav"}, \code{"mp3"}, or \code{"auto"}.}

\item{from}{Where to start reading in units.}

\item{to}{Where to stop reading.}

\item{units}{Units of \code{from} and \code{to} arguments.}
}
\value{
A list object.
}
\description{
This function cuts sound files by non-empty intervals on a specified tier. This tier commonly corresponds to the \code{main_tier} specified in \code{get_word_duration}.
}
\examples{
\dontrun{get_word_sound_files(sound_files = corpus_files[[1]], word_durations = wd_df, extra_tier = "ID", directory = "C:/Users/Project/chopped")}

}
\references{
Bořil, T., & Skarnitzl, R. (2016). Tools rPraat and mPraat. In P. Sojka, A. Horák, I. Kopeček, & K. Pala (Eds.), Text, Speech, and Dialogue (pp. 367-374). Springer International Publishing.
}
\author{
Dominic Schmitz
}
