% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CFBs.R
\name{plot_CFBs}
\alias{plot_CFBs}
\title{Compute correlations between columns of Mel-filterbank energy features}
\usage{
plot_CFBs(
  fbs,
  chunk_durations,
  file,
  tile_colour = "black",
  boundary_colour = "red",
  linewidth = 2
)
}
\arguments{
\item{fbs}{FBS list. Typically created with either \code{get_fbs} or \code{get_mel}.}

\item{chunk_durations}{Chunk duration list. Typically created with either \code{get_chunk_duration}.}

\item{file}{Name of the file you want to plot.}

\item{tile_colour}{Line colour of the tiles. Defaults to \code{"black"}.}

\item{boundary_colour}{Line colour of the boundary lines. Defaults to \code{"red"}.}

\item{linewidth}{Line width of the boundary lines. Defaults to \code{2}.}
}
\value{
A data frame.
}
\description{
This function computes correlations between columns of Mel-filterbank energy features. Note: Currently, this function does not take into account chunk boundaries.
}
\examples{
\dontrun{
plot_CFBs(fbs = fbs, chunk_durations = chunk_durations, file = "example.wav")
}

}
\references{
Wickham, H. (2016). ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York. ISBN 978-3-319-24277-4. Retrieved from https://ggplot2.tidyverse.org

Wickham, H. (2007). Reshaping Data with the {reshape} Package. Journal of Statistical Software, 21(12), 1-20. Retrieved from http://www.jstatsoft.org/v21/i12/
}
\author{
Dominic Schmitz
}
