% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphspn_to_wav.R
\name{sphspn_to_wav}
\alias{sphspn_to_wav}
\title{Convert SPH & SPN files to wav files}
\usage{
sphspn_to_wav(path)
}
\arguments{
\item{path}{The path to the directory containing SPH and SPN sound files.}
}
\value{
Nothing.
}
\description{
Using a bash script, this function converts SPH and SPN sound files to wav files.
Note: Currently, this only works on Windows. Additionally, \code{SoX - Sound eXchange} must be installed before running this function, see https://sourceforge.net/projects/sox/.
}
\examples{
\dontrun{sphspn_to_wav(path = "C:/Users/Dominic/Dropbox/HHU_2373/Compounds/CFBSFr/data")}

}
\author{
Dominic Schmitz
}
